+++ packages/pocket-ic/tests/tests.rs
@@ -237,6 +237,44 @@ fn test_create_canister_with_used_id_should_panic() {
 }
 
 #[test]
+fn test_cycle_scaling() {
+    let pic = PocketIcBuilder::new()
+        .with_application_subnet()
+        .with_fiduciary_subnet()
+        .build();
+    let app_canister_id =
+        pic.create_canister_on_subnet(None, None, pic.topology().get_app_subnets()[0]);
+    pic.add_cycles(app_canister_id, 100_000_000_000_000);
+    let fidu_canister_id =
+        pic.create_canister_on_subnet(None, None, pic.topology().get_fiduciary().unwrap());
+    pic.add_cycles(fidu_canister_id, 100_000_000_000_000);
+
+    let old_app_cycles = pic.cycle_balance(app_canister_id);
+    pic.install_canister(
+        app_canister_id,
+        UNIVERSAL_CANISTER_WASM.to_vec(),
+        vec![],
+        None,
+    );
+    let new_app_cycles = pic.cycle_balance(app_canister_id);
+    let app_cycles_delta = old_app_cycles - new_app_cycles;
+
+    let old_fidu_cycles = pic.cycle_balance(fidu_canister_id);
+    pic.install_canister(
+        fidu_canister_id,
+        UNIVERSAL_CANISTER_WASM.to_vec(),
+        vec![],
+        None,
+    );
+    let new_fidu_cycles = pic.cycle_balance(fidu_canister_id);
+    let fidu_cycles_delta = old_fidu_cycles - new_fidu_cycles;
+
+    // the fiduciary subnet has 28 nodes which is more than twice
+    // the number of nodes on an application subnet (13)
+    assert!(fidu_cycles_delta > 2 * app_cycles_delta);
+}
+
+#[test]
 fn test_random_subnet_selection() {
     // Application subnet has highest priority
     let pic = PocketIcBuilder::new()