+++ packages/pocket-ic/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleases
+
+### Added
+- Support cycle scaling according to subnet size.
 
 ## 2.0.0 - 2023-11-21